+++ src/test/NetCore/iPdfWriter.ConsoleAppCore60/Program.cs
@@ -96,11 +96,11 @@ namespace iPdfWriter
             logger.Info(" > Shows the use of add an enumerable (render as html) in a pdf document.");
             Sample16.Generate(logger, YesNo.No);
 
-            //// 17. Shows the use of add an enumerable (native render) in a pdf document.
-            //logger.Info("");
-            //logger.Info("> Start Pdf Sample 17");
-            //logger.Info(" > Shows the use of add an enumerable (native render) in a pdf document.");
-            //Sample17.Generate(logger, YesNo.No);
+            // 17. Shows the use of add an enumerable (native render) in a pdf document.
+            logger.Info("");
+            logger.Info("> Start Pdf Sample 17");
+            logger.Info(" > Shows the use of add an enumerable (native render) in a pdf document.");
+            Sample17.Generate(logger, YesNo.No);
 
             // 18. Shows the use of add a datatable (render as html) in a pdf document.
             logger.Info("");
@@ -108,11 +108,11 @@ namespace iPdfWriter
             logger.Info(" > Shows the use of add a datatable (render as html) in a pdf document.");
             Sample18.Generate(logger, YesNo.No);
 
-            //// 19. Shows the use of add a datatable (native render) in a pdf document.
-            //logger.Info("");
-            //logger.Info("> Start Pdf Sample 19");
-            //logger.Info(" > Shows the use of add a datatable (native render) in a pdf document.");
-            //Sample19.Generate(logger, YesNo.No);
+            // 19. Shows the use of add a datatable (native render) in a pdf document.
+            logger.Info("");
+            logger.Info("> Start Pdf Sample 19");
+            logger.Info(" > Shows the use of add a datatable (native render) in a pdf document.");
+            Sample19.Generate(logger, YesNo.No);
 
             // 20. Shows the use of add a datatable in a pdf document.
             logger.Info("");
@@ -125,6 +125,18 @@ namespace iPdfWriter
             logger.Info("> Start Pdf Sample 21");
             logger.Info(" > Shows the use of text and image replacement in a pdf document with custom font");
             Sample21.Generate(logger, YesNo.No);
+
+            // 22. Show how to extract pages from a pdf document.
+            logger.Info("");
+            logger.Info("> Start Pdf Sample 22");
+            logger.Info(" > Show how to extract pages from a pdf document");
+            Sample22.Generate(logger);
+
+            // 23. Show how to extract pages from a pdf document by search text
+            logger.Info("");
+            logger.Info("> Start Pdf Sample 23");
+            logger.Info(" > Show how to extract pages from a pdf document by search text");
+            Sample23.Generate();
 
             logger.Info("");
             logger.Debug(">End Logging<");